refactor: validator테스트를 도메인 테스트로 변경
