+++ src/test/java/lotto/validator/InputValidatorTest.java
@@ -14,12 +14,6 @@ class InputValidatorTest {
     private final String START_WITH_ZERO = "012," + ExceptionMessage.ERROR + ExceptionMessage.START_WItH_ZERO;
     private final String NO_ZERO = "0," + ExceptionMessage.ERROR + ExceptionMessage.ZERO_INPUT;
 
-    private final String OUT_OF_FORMAT =
-            "1,2,3,4,5,6,7," + ExceptionMessage.ERROR + ExceptionMessage.WINNING_FORMAT_ERROR;
-    private final String DUPLICATED_NUMBER =
-            "1,2,2,3,4,5," + ExceptionMessage.ERROR + ExceptionMessage.DUPLICATED_NUMBER;
-    private final String OUT_OF_BOUNDS = "1,2,3,4,46,5," + ExceptionMessage.ERROR + ExceptionMessage.OUT_OF_BOUNDS;
-
     @DisplayName("구매 금액 입력시 발생하는 모든 예외 검증 테스트")
     @ParameterizedTest
     @CsvSource({ONLY_NUMBER, START_WITH_ZERO, NO_ZERO})
@@ -28,18 +22,6 @@ class InputValidatorTest {
         // when
         // then
         assertThatThrownBy(() -> inputValidator.validateInputPrice(inputPrice))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(exceptionMessage);
-    }
-
-    @DisplayName("당첨 번호 입력시 발생할 수 있는 모든 예외 검증 테스트")
-    @ParameterizedTest
-    @CsvSource({OUT_OF_FORMAT, DUPLICATED_NUMBER, OUT_OF_BOUNDS})
-    void 잘못된_당첨번호_입력시_모든_예외_발생(String inputLottoNumber, String exceptionMessage) {
-        // given
-        // when
-        // then
-        assertThatThrownBy(() -> inputValidator.validateInputLottoNumber(inputLottoNumber))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(exceptionMessage);
     }
